模型管理对外接口
